Clean up converters and move them into new package
